requirements.txt
PyYAML~=6.0.2
pandas~=2.2.3
